Move autoregressive generation into ar module
